Tests/LibWeb/Text/expected/IntersectionObserver/explicit-root-non-descendant-should-not-intersect.txt

callback 1: isIntersecting=true, ratio=0.84375
total callbacks: 1
